feat: implementando facade para webhook

+++ src/main/java/br/selene/projectseleneback/presentation/dto/notification/RequestCheckoutNotificationDTO.java
@@ -4,7 +4,7 @@ import br.selene.projectseleneback.domain.checkout.PaymentCheckoutStatusEnum;
 
 public record RequestCheckoutNotificationDTO(
         String id,
-        int referenceId,
+        Long referenceId,
         PaymentCheckoutStatusEnum status
 ) {
 }
